+++ src/test/java/io/nagurea/smsupsdk/webhooks/retrieve/RetrieveWebhooksServiceIntTest.java
@@ -3,6 +3,8 @@ package io.nagurea.smsupsdk.webhooks.retrieve;
 import io.nagurea.smsupsdk.common.status.ResponseStatus;
 import io.nagurea.smsupsdk.webhooks.create.common.WebhookType;
 import io.nagurea.smsupsdk.webhooks.retrieve.response.Webhook;
+import io.nagurea.smsupsdk.webhooks.retrieve.response.list.RetrieveWebhooksResultResponse;
+import io.nagurea.smsupsdk.webhooks.retrieve.response.single.RetrieveSingleWebhookResultResponse;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
@@ -25,6 +27,8 @@ class RetrieveWebhooksServiceIntTest {
 
     private static final String YOUR_TOKEN = "Your Token";
     private static final String EXPECTED_TOKEN = "Bearer " + YOUR_TOKEN;
+
+    private static final String WEBHOOK_ID = "66";
 
     /**
      * Useless. Only here to see how services could be used with Spring
@@ -67,6 +71,28 @@ class RetrieveWebhooksServiceIntTest {
                         )
         );
 
+        mockServer.when(
+                request()
+                        .withPath("/webhook/" + WEBHOOK_ID)
+                        .withMethod("GET")
+                        .withHeader("Authorization", EXPECTED_TOKEN)
+        ).respond(
+                HttpResponse.response()
+                        .withStatusCode(200)
+                        .withBody(
+                                "{\n" +
+                                        "    \"status\": 1,\n" +
+                                        "    \"message\": \"OK\",\n" +
+                                        "    \"webhook\": {\n" +
+                                        "        \"webhook_id\": \"66\",\n" +
+                                        "        \"type\": \"DLR\",\n" +
+                                        "        \"url\": \"https://mywebhook.com/dlr\",\n" +
+                                        "        \"last_http_code\": 200\n" +
+                                        "    }\n" +
+                                        "}"
+                        )
+        );
+
     }
 
     @AfterAll
@@ -106,4 +132,32 @@ class RetrieveWebhooksServiceIntTest {
         assertEquals(expectedGetListsResponse, actualGetListsResponse);
 
     }
+
+    @Test
+    void retrieveWebhookById() throws IOException {
+        // given
+        RetrieveSingleWebhookResultResponse expectedGetListsResponse = RetrieveSingleWebhookResultResponse.builder()
+                .message("OK")
+                .responseStatus(ResponseStatus.OK)
+                .webhook(
+                        Webhook.builder()
+                                .id("66")
+                                .type(WebhookType.DLR)
+                                .url("https://mywebhook.com/dlr")
+                                .lastHttpCode(200)
+                                .build()
+                )
+                .build();
+
+
+        // when
+        final RetrieveSingleWebhookResultResponse actualGetListsResponse = retrieveWebhooksService.retrieveWebhookById(YOUR_TOKEN, WEBHOOK_ID).getEffectiveResponse();
+
+        // then
+        assertEquals(ResponseStatus.OK, actualGetListsResponse.getStatus());
+        assertEquals("OK", actualGetListsResponse.getMessage());
+        assertEquals(expectedGetListsResponse, actualGetListsResponse);
+
+    }
+
 }