+++ RELEASE-NOTES.md
@@ -75,6 +75,7 @@ Initial version with:
 
 ## [0.13.1](https://github.com/kisscool62/io.nagurea.smsupsdk/compare/v0.13.0...v0.13.1) (2022-08-16)
 * Webhooks > Retrieve a webhook
+* Webhooks > Retrieve a single webhook
 
 ### Upgrade Steps
 * [NO ACTION REQUIRED]
@@ -84,6 +85,7 @@ Initial version with:
 
 ### New Features
 * Webhooks > Retrieve a webhook
+* Webhooks > Retrieve a single webhook
 
 ### Bug Fixes